+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 ## master
 * Refactoring publication and configuration logic
 * Bumping dependencies and migration to Kotlin 1.9.10
+* Cleaning up in favor of new Kotlin/Native memory model
+* `TestMode` is deprecated now since we don't need to run on Local or Multithreaded modes anymore
+* Removing stately isolate and replacing with `stately-concurrent-collections`
 
 ## 2.12.1
 * Fixed publication